+++ packages/web-component/src/lib/helpers/templates.ts
@@ -3,7 +3,7 @@ import {
   DESCOPE_ATTRIBUTE_EXCLUDE_FIELD,
   HAS_DYNAMIC_VALUES_ATTR_NAME,
 } from '../constants';
-import { ScreenState } from '../types';
+import { ComponentsConfig, ScreenState } from '../types';
 
 const replaceElementMessage = (
   baseEle: DocumentFragment,
@@ -103,6 +103,40 @@ const replaceProvisionURL = (
   });
 };
 
+const setElementConfig = (
+  baseEle: DocumentFragment,
+  componentsConfig: ComponentsConfig,
+  logger?: { error: (message: string, description: string) => void },
+) => {
+  if (!componentsConfig) {
+    return;
+  }
+  // collect components that needs configuration from DOM
+  Object.keys(componentsConfig).forEach((componentName) => {
+    baseEle.querySelectorAll(`[name=${componentName}]`).forEach((comp) => {
+      const config = componentsConfig[componentName];
+
+      Object.keys(config).forEach((attr) => {
+        let value = config[attr];
+
+        if (typeof value !== 'string') {
+          try {
+            value = JSON.stringify(value);
+          } catch (e) {
+            logger.error(
+              `Could not stringify value "${value}" for "${attr}"`,
+              e.message,
+            );
+            value = '';
+          }
+        }
+
+        comp.setAttribute(attr, value);
+      });
+    });
+  });
+};
+
 /**
  * Update a screen template based on the screen state
  *  - Show/hide error messages
@@ -111,6 +145,7 @@ const replaceProvisionURL = (
 export const updateTemplateFromScreenState = (
   baseEle: DocumentFragment,
   screenState?: ScreenState,
+  componentsConfig?: ComponentsConfig,
   errorTransformer?: (error: { text: string; type: string }) => string,
   logger?: { error: (message: string, description: string) => void },
 ) => {
@@ -127,6 +162,7 @@ export const updateTemplateFromScreenState = (
   replaceElementMessage(baseEle, 'error-message', errorText);
   replaceProvisionURL(baseEle, screenState?.totp?.provisionUrl);
   replaceElementTemplates(baseEle, screenState);
+  setElementConfig(baseEle, componentsConfig, logger);
   replaceTemplateDynamicAttrValues(baseEle, screenState);
 };